Clean up some more
